Modified test_olmo_causal.py to be exactly like test_hf_causal.py and updated olmo.py to load full float32 model for causal tasks

+++ deepchem/models/torch_models/olmo.py
@@ -45,8 +45,7 @@ class Olmo(HuggingFaceModel):
         if isinstance(model, str):
             model_path = model
             if task_type == "causal_lm":
-                model = OlmoForCausalLM.from_pretrained(
-                    model_path, torch_dtype=torch.bfloat16)
+                model = OlmoForCausalLM.from_pretrained(model_path)
             else:
                 model = OlmoModel.from_pretrained(model_path,
                                                   torch_dtype=torch.bfloat16)
